Changes to the structure of the Project
Passed from in memory design to a connection to a database (Postgres), Added new types for new endpoints that will be used in the main project. TODOS, start developing the front-end to test the endpoints created. Implement the functionality of the API from scracth, only using main libraries or, if necessary mux. Middleware will only be focused when the previous steps are completed.